feat(api) : exception moved to server
