+++ packages/api/src/server/exceptions/exception.abstract.ts
