+++ packages/api/src/server/exceptions/exceptions.module.ts
